added navbar changed header changed general page layout fixed footer adjusted flexbox and grid
