+++ src/components/Header/Header.tsx
@@ -1,8 +1,6 @@
 import React, { FC } from 'react';
-import Link from '../Link';
 
 import './Header.scss';
-import brand from '../../img/logo-mini.png';
 
 type Props = JSX.IntrinsicElements['header'] & {
   path?: string;
@@ -15,26 +13,7 @@ const Header: FC<Props> = ({ className, path }) => {
     ? `${baseStyle} ${headerStyle}  ${className}`
     : `${baseStyle} ${headerStyle}`;
 
-  return (
-    <header className={classNames}>
-      <div className="section-1">
-        <Link to="/">
-          <img src={brand} alt="brand" className="brand" />
-        </Link>
-      </div>
-      <div className="section-2">
-        <Link to="/login">Wiki</Link>
-        <Link to="/login">Lore</Link>
-        <Link to="/login">Store</Link>
-      </div>
-      <div className="section-3">
-        <Link to="/login">Login</Link>
-        <Link cta to="/feed">
-          Register
-        </Link>
-      </div>
-    </header>
-  );
+  return <header className={classNames} />;
 };
 
 export default Header;
