+++ src/components/Navigation/Navigation.scss
@@ -1,0 +1,35 @@
+@import '../../styles/colors';
+
+/*--------------------------------------------------------------
+General styles
+--------------------------------------------------------------*/
+
+.navContainer {
+  padding: 1em;
+  background-color: $color-black-dark;
+  min-height: 60px;
+  min-width: 150px;
+  display: inline-flex;
+  justify-content: space-between;
+}
+
+div img {
+  display: inline;
+  min-height: 50px;
+  width: 150px;
+  margin: 0 auto;
+}
+
+a {
+  font-size: 21px;
+  margin-left: 9px;
+  margin-right: 9px;
+  color: $color-secondary-light;
+}
+
+a.active {
+  font-size: 21px;
+  color: $color-secondary-light;
+  margin-left: 9px;
+  margin-right: 9px;
+}
