+++ src/components/Navigation/Navigation.stories.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import Navigation from './Navigation';
+
+storiesOf('Navigation', module).add('Default', () => <Navigation />);
