+++ src/routes/Home/Home.scss
@@ -1,33 +1,35 @@
 @import '../../styles/colors';
+@import '../../styles/index';
 
-.home .top {
-  background: url('https://thetrove.is/images/header-bg/rotfm.jpg') no-repeat center/cover;
-  background-attachment: fixed;
-  height: 60vh;
+#main {
+  background-color: $color-black-dark;
+  border-bottom: 3px solid $color-secondary-dark;
+  min-height: 400px;
+  flex: 1;
+  flex-direction: column;
+  position: relative;
   justify-content: center;
-  align-items: center;
 }
 
-.home .logo {
-  max-width: 1000px;
-  display: block;
-  margin: 0 auto;
-}
-.home .main {
+#bottom {
   background-color: $color-black-dark;
-  border-bottom: 1px solid $color-secondary;
+  min-height: 400px;
+  flex: 1;
+  flex-direction: column;
+  position: relative;
+  justify-content: center;
 }
-.home .bottom {
-  background-color: $color-black-dark;
+
+p {
+  color: white;
 }
 
 /*--------------------------------------------------------------
 If no grid support, limit width to 50em and center align
 --------------------------------------------------------------*/
 @supports not (display: grid) {
-  .home .top,
-  .home .main,
-  .home .bottom {
+  .home #main,
+  .home #bottom {
     max-width: 50em;
     margin: 0 auto;
   }
@@ -38,27 +40,27 @@ CSS Grid layout for wider screens, when browser supports grid:
 --------------------------------------------------------------*/
 
 @supports (display: grid) {
-  .home .top {
-    grid-area: home-top;
-  }
-
-  .home .main {
+  .home #main {
     grid-area: home-main;
   }
 
-  .home .bottom {
+  .home #bottom {
     grid-area: home-bottom;
   }
 
   @media screen and (min-width: 800px) {
     .home {
       display: grid;
-      grid-template-columns: 3fr 1fr;
-      grid-template-rows: 1fr 1fr 1fr;
+      position: absolute;
+      top: 0;
+      bottom: 0;
+      left: 0;
+      right: 0;
+      grid-template-columns: 1fr 1fr 1fr 1fr;
+      grid-template-rows: 1fr 1fr;
       grid-template-areas:
-        'home-top home-top'
-        'home-main home-main'
-        'home-bottom home-bottom';
+        'home-main home-main home-main home-main'
+        'home-bottom home-bottom home-bottom home-bottom';
     }
   }
 }
