+++ src/routes/Home/Home.tsx
@@ -1,16 +1,29 @@
 import React, { FunctionComponent } from 'react';
+import { useLocation } from 'react-router-dom';
 
+import Header from '../../components/Header';
 import './Home.scss';
-import logo from '../../img/logo.png';
 
-const Home: FunctionComponent = () => (
-  <div className="grid home">
-    <h1 className="top flex-container column">
-      <img src={logo} alt="logo" className="logo" />
-    </h1>
-    <div className="main">Main</div>
-    <div className="bottom">bottom</div>
-  </div>
-);
+const Home: FunctionComponent = () => {
+  const location = useLocation();
+  const path = location.pathname.split('/')[1];
+
+  return (
+    <div id="container" className="flex-container column">
+      <Header path={path} className="flex-item" />
+
+      <div id="main" className="flex-item item ">
+        <div className="grid home">
+          <p>ha</p>
+        </div>
+      </div>
+      <div id="bottom" className="flex-item item">
+        <div className="grid home">
+          <p>ha</p>
+        </div>
+      </div>
+    </div>
+  );
+};
 
 export default Home;
