+++ src/styles/index.scss
@@ -11,6 +11,12 @@ html {
 
 body {
   margin: 0;
+  box-sizing: border-box;
+  background-color: $color-black-dark;
+}
+
+* {
+  box-sizing: inherit;
 }
 
 img {
@@ -62,9 +68,7 @@ body,
 button,
 input,
 select,
-textarea,
-a {
-  color: $color-black-dark;
+textarea {
   font-family: 'Exo', sans-serif;
   font-size: 18px;
   line-height: 1.5;
@@ -88,6 +92,7 @@ General styles
 
 .fullpage {
   min-height: 100vh;
+  position: relative;
 }
 
 .main-content {
@@ -129,21 +134,15 @@ CSS Flexbox layout for wider screens, when browser supports flexbox:
     flex-direction: column;
   }
 
-  .flex-item {
+  .flex-item.item {
     display: inherit;
     -webkit-align-items: center;
     -ms-flex-align: center;
     align-items: center;
   }
 
-  .flex-item p {
-    width: 100%;
-    text-align: center;
-    color: $color-white;
-  }
-
   //use a second flex in main-content to force max available height
-  .main-content {
+  .main-content.content {
     display: flex;
     -webkit-flex-direction: column;
     -ms-flex-direction: column;
